fix(types): close mypy in validators/rubric (mypy -32)
Typing-only fixes across the rubric checkers, no behavior change:
- regex_patterns params defaulted to None but were annotated dict -> widen to
  dict | None (bodies already guard `if regex_patterns`).
- Wrap Any-typed returns to their declared bool/str/float, cast the embedding
  matrix return where the callable is untyped.
- section1 rx_* attrs come from dict.get() (Any | None) but are never None in
  practice -> annotate Any so the .search() calls type-check.
- similarity: annotate the [None]*n vector buffer as list[Sequence[float] | None]
  and cast the fully-populated return; drop redundant re-annotations of `scores`
  in the fallback branches (no-redef).
- toc: heading_accurate is bool | None (None sentinel before the decision).
- section3: bind items[-1].details (dict | None) to a local before the f-string.
- document_utils: rename the reused loop var so the Optional lookup no longer
  clashes with the non-optional ReadmeSection.
- scorer: embedding_function is Optional (falls back to None); import the type.

mypy src 150 -> 118. ruff clean, 268 validator/rubric tests green.

+++ src/content_factory/generation/validators/rubric/utils.py
@@ -126,7 +126,7 @@ def semantic_similarity(text1: str, text2: str, lang: str = "ru", embedding_func
             norm_a = sqrt(sum(a * a for a in va))
             norm_b = sqrt(sum(b * b for b in vb))
             if norm_a > 0 and norm_b > 0:
-                return dot_product / (norm_a * norm_b)
+                return float(dot_product / (norm_a * norm_b))
             else:
                 safe_print(f"[SEMANTIC] Нулевая норма векторов: norm_a={norm_a}, norm_b={norm_b}", flush=True)
                 return 0.0
